refactor executors
