+++ pkg/executor/k6/k6_parser_test.go
