+++ pkg/executor/k6/k6tests/fail.js
