+++ pkg/executor/playwright/parser_test.go
